Some update with key-value-body form and notify functions.
The key-value-body is like:

  :key value
  :key value
  body

which is used to constract the iter macros with parameters controlled.

The parser is defined in macro-helper.lisp.

The notify function I found very useful especially if you want to
do some heavy computations and want to hang out for other tasks,
but stil want to get notified. That's it. Computational Physics
is a little time killer.